Add sections on EM algorithm and GMMs with EM
This commit expands the beamer document by adding a section on the EM
algorithm and a brief section that lays out how to approach Gaussian
Mixture Models with the EM algorithm. I am an idiot and forgot to
include all files in my last commit, which is why there are two with the
same name so quickly after one another.